Tick Rate Change - Change Tick Rate from a double to an int. (we were never using fractions of a tick in this situation to begin with) - It could be an issue within the TimerRegistry itself. Where it has a slight delay in checking and some ticks don't process on time. https://github.com/TrueUO/TrueUO/blob/master/Scripts/Spells/Base/SpellHelper.cs#L1720

+++ Scripts/Spells/Base/SpellHelper.cs
@@ -1756,7 +1756,7 @@ namespace Server.Spells
         int PoisResistOffset { get; }
         int NrgyResistOffset { get; }
 
-        double TickRate { get; }
+        int TickRate { get; }
         void OnTick(Mobile m);
 
         void DoEffect(Mobile m);
